+++ nodes/CAD/straight_skeleton_2d_offset.py
@@ -27,40 +27,21 @@ from sverchok.nodes.analyzer.mesh_filter import Edges
 from sverchok.nodes.vector.vertices_sort import sort_vertices_by_connexions
 from sverchok.utils.modules.polygon_utils import areas_from_polygons
 from sverchok.utils.sv_operator_mixins import SvGenericNodeLocator
-#from .stright_skeleton_2d_extrude import create_list2_in_range
+from .straight_skeleton_2d_extrude import create_list2_in_range
+#import networkx as nx
+
+from time import time
 
 enable_module = False
 try:
     from more_itertools import sort_together
     import pySVCGAL
-    from pySVCGAL.pySVCGAL import pySVCGAL_straight_skeleton_2d_extrude
+    from pySVCGAL.pySVCGAL import pySVCGAL_straight_skeleton_2d_offset
     enable_module = True
 except ModuleNotFoundError:
     enable_module = False
 
 
-# if values out of range then values will be aligned to range
-# Lst value will repeat until end
-def create_list2_in_range(length, values_in, ss_shapes_modes):
-    objects_mask = []
-    min_allowed_value = min(ss_shapes_modes)
-    max_allowed_value = max(ss_shapes_modes)
-    for I in range(length):
-        if I<len(values_in):
-            val = values_in[I]
-        else:
-            val = values_in[-1]
-        
-        if val < min_allowed_value:
-            val = min_allowed_value
-        elif max_allowed_value < val:
-            val = max_allowed_value
-        
-        objects_mask.append( val )
-        pass
-
-    res = objects_mask
-    return res
 
 def vertices_sort_by_edges(verts_in, edges_in):
     edges_indexes = list(itertools.chain(*edges_in))
@@ -116,6 +97,25 @@ def vertices_sort_by_edges(verts_in, edges_in):
     return verts_out
     pass
 
+# Сюда можно передать только контуры у которых edges boundary=True
+# def separate_edges(verts_in, edges_in):
+#     np_poly_edge_in = np.array(edges_in)
+#     np_verts = np.array(verts_in)
+#     G = nx.Graph()
+#     G.add_edges_from(edges_in)
+#     contours = []
+#     verts = []
+#     for c in nx.connected_components(G):
+#         print( len(G.subgraph(c)))
+#         #print( list(G.subgraph(c)))
+#         sg0 = G.subgraph(c)
+#         contour1_indexes = nx.cycle_basis(sg0)[0]  # get cycle. https://networkx.org/documentation/stable/reference/algorithms/generated/networkx.algorithms.cycles.cycle_basis.html#networkx.algorithms.cycles.cycle_basis
+#         vert_c1 = np_verts[contour1_indexes]
+#         contours.append(contour1_indexes)
+#         verts.append(vert_c1)
+#     return verts
+#     pass
+
 def separate_loose_mesh(verts_in, poly_edge_in):
         ''' separate a mesh by loose parts.
         input:
@@ -126,6 +126,66 @@ def separate_loose_mesh(verts_in, poly_edge_in):
           2. separated list of edges/polygons with new indices of separated elements
           3. separated list of edges/polygons (like 2) with old indices
         '''
+        
+        # # # def dict_key_of_value(d, val):
+        # # #     res = None
+        # # #     for KV in d:
+        # # #         if val in d[KV]:
+        # # #             res =  KV
+        # # #     return res
+        
+        # # # np_poly_edge_in = np.array(poly_edge_in)
+        # # # dict_objects_indexes = dict()
+        # # # for I in range(len(poly_edge_in)):
+            
+        # # #     object_I_owner = None
+        # # #     for IJ in range(I+1, len(poly_edge_in)):
+        # # #         if np.any(np.intersect1d(poly_edge_in[I], poly_edge_in[IJ])):
+
+        # # #             object_I_owner = dict_key_of_value(dict_objects_indexes, I)
+        # # #             if object_I_owner is None:
+        # # #                 object_I_owner = dict_key_of_value(dict_objects_indexes, IJ)
+        # # #                 if object_I_owner is None:
+        # # #                     if len(dict_objects_indexes)==0:
+        # # #                         object_I_owner = 0
+        # # #                     else:
+        # # #                         object_I_owner = max( list(dict_objects_indexes.keys()) )
+        # # #                         object_I_owner+=1
+        # # #                     dict_objects_indexes[object_I_owner] = set()
+
+        # # #             dict_objects_indexes[object_I_owner].update([I, IJ])
+        # # #         # else:
+        # # #         #     if (poly_edge_in[I] in np_poly_edge_in[IJ+1:])==False:
+        # # #         #         break
+        # # #         #     pass
+
+        # # #     if object_I_owner is None:
+        # # #         if dict_key_of_value(dict_objects_indexes, I) is None:
+        # # #             if len(dict_objects_indexes)==0:
+        # # #                 object_I_owner = 0
+        # # #             else:
+        # # #                 object_I_owner = max( list(dict_objects_indexes.keys()) )
+        # # #                 object_I_owner+=1
+        # # #             dict_objects_indexes[object_I_owner] = set([I])
+        # # #             #dict_objects_indexes[object_I_owner].update([I])
+        # # #             pass
+        # # #         pass
+        # # #     pass
+        # # # pass
+
+        # # # objects_sets = []
+        # # # for KV in dict_objects_indexes:
+        # # #     objectI = list(dict_objects_indexes[KV])
+        # # #     idx_faces = []
+        # # #     for idx in objectI:
+        # # #         idx_faces.extend(poly_edge_in[idx])
+        # # #     objects_sets.append(idx_faces)
+        # # # pass
+        # # # node_set_list = objects_sets
+
+        ########################
+
+
         verts_out = []
         poly_edge_out = []
         poly_edge_old_indexes_out = []  # faces with old indices 
@@ -185,16 +245,30 @@ def ShowMessageBox(message = "", title = "Message Box", icon = 'INFO'):
         self.layout.label(text=message)
     bpy.context.window_manager.popup_menu(draw, title = title, icon = icon)
 
-# Привести параметры и интерфейс в соответствии с CGAL
-class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.types.Node):
+
+# # Operator to save data in .dat format file for test in CGAL (Hidden in production)
+# class SvSaveCGALDatFile(bpy.types.Operator, SvGenericNodeLocator):
+#     ''' Save coords and angles to the file .dat for CGAL '''
+#     bl_idname = "node.sverchok_save_cgal_dat_file"
+#     bl_label = "Save coords and angles to the file .dat for CGAL"
+    
+#     def sv_execute(self, context, node):
+#         if hasattr(node, 'saveCGALDatFile')==True:
+#             node.saveCGALDatFile()
+#             #text = node.dataAsString()
+#             #context.window_manager.clipboard = text
+#             ShowMessageBox("File saved")
+#         pass
+
+class SvStraightSkeleton2DOffset(ModifierLiteNode, SverchCustomTreeNode, bpy.types.Node):
     """
     Triggers: Merge two 2d meshes
 
     Each mesh can have disjoint parts
     Only X and Y coordinate takes in account
     """
-    bl_idname = 'SvStraightSkeleton2DExtrude'
-    bl_label = 'Straight Skeleton 2D Extrude (Alpha)'
+    bl_idname = 'SvStraightSkeleton2DOffset'
+    bl_label = 'Straight Skeleton 2D Offset (Alpha)'
     bl_icon = 'MOD_OUTLINE'
 
     sv_dependencies = ['pySVCGAL', 'more_itertools']
@@ -215,64 +289,46 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
         op.tree_name = self.id_data.name
         return op
 
-    ss_angles: FloatProperty(
-        default=0.785398, name="Taper angle", update=updateNode,
-        description = "Taper angle",
-        subtype='ANGLE',
+    offset_modes = [
+            ('OBJECT_ALL_OFFSETS', "All", "Every object get all offsets", 'THREE_DOTS', 0),
+            ('OBJECT_ONE_OFFSET' , "One", "Every object get one offset", 'DECORATE', 1),
+        ]
+    offset_mode : EnumProperty(
+        name = "Offset mode",
+        default = 'OBJECT_ALL_OFFSETS',
+        description = "How many offsets per object (One or All)",
+        items = offset_modes,
+        update = updateNode
+        ) # type: ignore
+
+    ss_offset1: FloatProperty(
+        name="Offsets   ",
+        default=0.1,
+        description = "Offsets",
+        update=updateNode,
+        #subtype='DISTANCE',
     ) # type: ignore
 
-    ss_height: FloatProperty(
-        default=1.0, name="Max height", update=updateNode,
-        description = "Max height. Disabled if (socket connected or Restrict height is off)",
+
+    altitude_modes = [
+            ('OBJECT_ALL_ALTITUDES', "All", "Every object get all altitudes", 'THREE_DOTS', 0),
+            ('OBJECT_ONE_ALTITUDE' , "One", "Every object get one altitude", 'DECORATE', 1),
+        ]
+    altitude_mode : EnumProperty(
+        name = "Altitudes",
+        default = 'OBJECT_ALL_ALTITUDES',
+        description = "How many Altitudes per object (One or All)",
+        items = altitude_modes,
+        update = updateNode
+        ) # type: ignore
+
+
+    ss_altitude1: FloatProperty(
+        name="Altitudes",
+        default=1, 
+        description = "Altitude of offsets",
+        update=updateNode,
     ) # type: ignore
-
-    ss_restrict_height1: BoolProperty(
-        name="Restrict height",
-        description='Restrict height of object. (If no then all heights wil be unlimited height)',
-        default=True, update=updateNode) # type: ignore
-
-
-    # offset_modes = [
-    #         ('OBJECT_ALL_OFFSETS', "All", "Every object get all offsets", 'THREE_DOTS', 0),
-    #         ('OBJECT_ONE_OFFSET' , "One", "Every object get one offset", 'DECORATE', 1),
-    #     ]
-    # offset_mode : EnumProperty(
-    #     name = "Offset mode",
-    #     default = 'OBJECT_ALL_OFFSETS',
-    #     description = "How many offsets per object (One or All)",
-    #     items = offset_modes,
-    #     update = updateNode
-    #     ) # type: ignore
-
-    # ss_offset1: FloatProperty(
-    #     name="Offsets   ",
-    #     default=0.1,
-    #     description = "Offsets",
-    #     update=updateNode,
-    #     #subtype='DISTANCE',
-    # ) # type: ignore
-
-
-    # altitude_modes = [
-    #         ('OBJECT_ALL_ALTITUDES', "All", "Every object get all altitudes", 'THREE_DOTS', 0),
-    #         ('OBJECT_ONE_ALTITUDE' , "One", "Every object get one altitude", 'DECORATE', 1),
-    #     ]
-    # altitude_mode : EnumProperty(
-    #     name = "Altitudes",
-    #     default = 'OBJECT_ALL_ALTITUDES',
-    #     description = "How many Altitudes per object (One or All)",
-    #     items = altitude_modes,
-    #     update = updateNode
-    #     ) # type: ignore
-
-
-    # ss_altitude1: FloatProperty(
-    #     name="Altitudes",
-    #     default=1, 
-    #     description = "Altitude of offsets",
-    #     update=updateNode,
-    #     #subtype='DISTANCE',
-    # ) # type: ignore
 
     ss_shapes_modes = [
             ( 'ORIGINAL_BOUNDARIES',          "Original", "Offset of original boundaries", 'RENDER_ANIMATION', 0),
@@ -304,33 +360,40 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
         description='Show additional debug info in console',
         default=True, update=updateNode) # type: ignore
 
-    # contours_or_mesh: BoolProperty(
-    #     name='Ret',
-    #     description='off - contours, on - mesh',
-    #     default=True, update=updateNode) # type: ignore
-
-    join_modes = [
+    source_objects_join_modes = [
             ('SPLIT', "Split", "Separate the result meshes into individual meshes", 'SNAP_VERTEX', 0),
-            ('KEEP' , "Keep", "Keep as source meshes", 'SYNTAX_ON', 1),
-            ('MERGE', "Merge", "Join all results meshes into a single mesh", 'STICKY_UVS_LOC', 2)
+            ('KEEP' , "Keep", "Keep as input meshes", 'SYNTAX_ON', 1),
+            ('MERGE', "Merge", "Join all meshes into a single mesh", 'STICKY_UVS_LOC', 2)
         ]
 
-    join_mode : EnumProperty(
-        name = "Output mode",
-        items = join_modes,
+    source_objects_join_mode : EnumProperty(
+        name = "How process input objects",
+        items = source_objects_join_modes,
         default = 'KEEP',
         update = updateNode) # type: ignore
 
-    # res_types = [
-    #         ('CONTOURS', "Contours", "Edges of contours", 'SNAP_VERTEX', 0),
-    #         ('FACES' , "Faces", "Fill faces", 'SYNTAX_ON', 1),
-    #     ]
+    results_join_modes = [
+            ('SPLIT', "Split", "Separate source meshes into individual meshes", 'SNAP_VERTEX', 0),
+            ('KEEP' , "Keep", "Keep as source meshes", 'SYNTAX_ON', 1),
+            ('MERGE', "Merge", "Join all source meshes into a single mesh (if you need general external contour)", 'STICKY_UVS_LOC', 2)
+        ]
 
-    # res_type : EnumProperty(
-    #     name = "Result",
-    #     items = res_types,
-    #     default = 'FACES',
-    #     update = updateNode) # type: ignore
+    results_join_mode : EnumProperty(
+        name = "Output mode",
+        items = results_join_modes,
+        default = 'KEEP',
+        update = updateNode) # type: ignore
+
+    res_types = [
+            ('CONTOURS', "Contours", "Edges of contours", 'SNAP_VERTEX', 0),
+            ('FACES' , "Faces", "Fill faces", 'SYNTAX_ON', 1),
+        ]
+
+    res_type : EnumProperty(
+        name = "Result",
+        items = res_types,
+        default = 'CONTOURS',
+        update = updateNode) # type: ignore
 
     objects_mask_modes = [
             ('BOOLEANS', "Booleans", "Boolean values (0/1) as mask of Voronoi Sites per objects [[0,1,0,0,1,1],[1,1,0,0,1],...]. Has no influence if socket is not connected (All sites are used)", 0),
@@ -348,19 +411,16 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
         description="Invert mask of sites. Has no influence if socket is not connected (All sites are used)",
         update = updateNode) # type: ignore
 
-    angles_modes = [
-            ('OBJECTS', "Objects", "Use this angle on all edges of one objects", 'OBJECT_DATA', 0),
-            ('EDGES', "Edges", "Use many angles for objects", 'MOD_EDGESPLIT', 1),
-        ]
-    angles_mode : EnumProperty(
-        name = "Mask of Objects",
-        items = angles_modes,
-        default = 'OBJECTS',
-        update = updateNode
-        ) # type: ignore
+    def draw_vertices_in_socket(self, socket, context, layout):
+        if socket.is_linked:  # linked INPUT or OUTPUT
+            layout.label(text=f"{socket.label}. {socket.objects_number or ''}")
+        else:
+            layout.label(text=f'{socket.label}')
+        layout.prop(self, 'source_objects_join_mode', text='')
+        pass
 
     def draw_vertices_out_socket(self, socket, context, layout):
-        layout.prop(self, 'join_mode', text='')
+        layout.prop(self, 'results_join_mode', text='')
         if socket.is_linked:  # linked INPUT or OUTPUT
             layout.label(text=f"{socket.label}. {socket.objects_number or ''}")
         else:
@@ -396,56 +456,6 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
             col.prop(self, 'ss_shapes_mode1', text='Shapes mode')
         pass
 
-    def draw_ss_restrict_heights_in_socket(self, socket, context, layout):
-        grid = layout.grid_flow(row_major=False, columns=3, align=True)
-        col = grid.row(align=True)
-        if socket.is_linked:
-            socket_label = socket.objects_number if hasattr(socket, "objects_number")==True else '-'
-            col.label(text=f"Restrict heights {socket_label}")
-        else:
-            col.prop(self, 'ss_restrict_height1', text='Restrict heights')
-        pass
-
-    def draw_angles_mode_in_socket(self, socket, context, layout):
-        grid = layout.grid_flow(row_major=False, columns=3, align=True)
-        col = grid.column()
-        col.prop(self, 'ss_angles')
-        if socket.is_linked==True:
-            col.enabled = False
-        else:
-            col.enabled = True
-        # Временно отключено
-        #grid.prop(self, 'angles_mode', expand=True, icon_only=True) 
-
-    def draw_ss_height_in_socket(self, socket, context, layout):
-        grid = layout.grid_flow(row_major=False, columns=3, align=True)
-        col = grid.column()
-        col.prop(self, 'ss_height')
-        col.enabled = True
-        if socket.is_linked==True:
-            col.enabled = False
-        pass
-
-    # def draw_offset_mode_in_socket(self, socket, context, layout):
-    #     grid = layout.grid_flow(row_major=False, columns=3, align=True)
-    #     col = grid.column() # align=True
-    #     col.prop(self, 'ss_offset1')
-    #     if socket.is_linked==True:
-    #         col.enabled = False
-    #     else:
-    #         col.enabled = True
-    #     grid.prop(self, 'offset_mode', expand=True, icon_only=True) 
-
-    # def draw_altitude_mode_in_socket(self, socket, context, layout):
-    #     grid = layout.grid_flow(row_major=False, columns=3, align=True)
-    #     col = grid.column()
-    #     col.prop(self, 'ss_altitude1')
-    #     if socket.is_linked==True:
-    #         col.enabled = False
-    #     else:
-    #         col.enabled = True
-    #     grid.prop(self, 'altitude_mode', expand=True, icon_only=True) 
-
     def draw_objects_mask_in_socket(self, socket, context, layout):
         grid = layout.grid_flow(row_major=True, columns=2)
         if not socket.is_linked:
@@ -463,9 +473,29 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
         col3 = grid.column()
         col3.prop(self, "objects_mask_mode", expand=True)
 
+    def draw_offset_mode_in_socket(self, socket, context, layout):
+        grid = layout.grid_flow(row_major=False, columns=3, align=True)
+        col = grid.column() # align=True
+        col.prop(self, 'ss_offset1')
+        if socket.is_linked==True:
+            col.enabled = False
+        else:
+            col.enabled = True
+        grid.prop(self, 'offset_mode', expand=True, icon_only=True) 
+
+    def draw_altitude_mode_in_socket(self, socket, context, layout):
+        grid = layout.grid_flow(row_major=False, columns=3, align=True)
+        col = grid.column()
+        col.prop(self, 'ss_altitude1')
+        if socket.is_linked==True:
+            col.enabled = False
+        else:
+            col.enabled = True
+        grid.prop(self, 'altitude_mode', expand=True, icon_only=True) 
+
     def draw_buttons(self, context, layout):
         col = layout.column()
-        #col.row(align=True).prop(self, 'res_type', expand=True)
+        col.row(align=True).prop(self, 'res_type', expand=True)
         col.prop(self, 'only_tests_for_valid')
         col.prop(self, 'force_z_zero')
         col.prop(self, 'verbose_messages_while_process') 
@@ -481,39 +511,26 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
 
     def sv_init(self, context):
 
-        self.width = 230
+        self.width = 220
 
         self.inputs.new('SvVerticesSocket', 'vertices')
         self.inputs.new('SvStringsSocket' , 'edges')
         self.inputs.new('SvStringsSocket' , 'polygons')
         self.inputs.new('SvStringsSocket' , 'ss_shapes_modes')
-
-        self.inputs.new('SvStringsSocket' , 'ss_angles').prop_name = 'ss_angles'
-        self.inputs.new('SvStringsSocket' , 'ss_restrict_heights').prop_name = 'ss_restrict_height1'
-        self.inputs.new('SvStringsSocket' , 'ss_height').prop_name = 'ss_height'
-
-        # self.inputs.new('SvStringsSocket' , 'ss_offsets').prop_name = 'ss_offset1'
-        # self.inputs.new('SvStringsSocket' , 'ss_altitudes').prop_name = 'ss_altitude1'
+        self.inputs.new('SvStringsSocket' , 'ss_offsets').prop_name = 'ss_offset1'
+        self.inputs.new('SvStringsSocket' , 'ss_altitudes').prop_name = 'ss_altitude1'
         self.inputs.new('SvStringsSocket' , 'objects_mask').label = "Mask of Objects"
         self.inputs.new('SvTextSocket'    , 'file_name')
 
         self.inputs['vertices'].label = 'Vertices'
+        self.inputs['vertices'].custom_draw = 'draw_vertices_in_socket'
         self.inputs['edges'].label = 'Edges'
         self.inputs['polygons'].label = 'Polygons'
         self.inputs['ss_shapes_modes'].custom_draw = 'draw_ss_shapes_modes_in_socket'
-
-        self.inputs['ss_angles'].label = 'Angles'
-        self.inputs['ss_angles'].custom_draw = 'draw_angles_mode_in_socket'
-        self.inputs['ss_restrict_heights'].label = 'Restricts Heights'
-        self.inputs['ss_restrict_heights'].custom_draw = 'draw_ss_restrict_heights_in_socket'
-        self.inputs['ss_height'].label = 'Height'
-        self.inputs['ss_height'].custom_draw = 'draw_ss_height_in_socket'
-
-
-        # self.inputs['ss_offsets'].label = 'Offsets'
-        # self.inputs['ss_offsets'].custom_draw = 'draw_offset_mode_in_socket'
-        # self.inputs['ss_altitudes'].label = 'Offsets'
-        # self.inputs['ss_altitudes'].custom_draw = 'draw_altitude_mode_in_socket'
+        self.inputs['ss_offsets'].label = 'Offsets'
+        self.inputs['ss_offsets'].custom_draw = 'draw_offset_mode_in_socket'
+        self.inputs['ss_altitudes'].label = 'Offsets'
+        self.inputs['ss_altitudes'].custom_draw = 'draw_altitude_mode_in_socket'
         self.inputs['objects_mask'].custom_draw = 'draw_objects_mask_in_socket'
         self.inputs['file_name'].label = 'File Name'
         self.inputs['file_name'].hide = True
@@ -521,16 +538,16 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
         self.outputs.new('SvVerticesSocket', 'vertices')
         self.outputs.new('SvStringsSocket' , 'edges')
         self.outputs.new('SvStringsSocket' , 'polygons')
-        # self.outputs.new('SvStringsSocket' , 'offsets')
-        # self.outputs.new('SvStringsSocket' , 'altitudes')
+        #self.outputs.new('SvStringsSocket' , 'offsets')
+        #self.outputs.new('SvStringsSocket' , 'altitudes')
         self.outputs.new('SvVerticesSocket', 'failed_contours_vertices')
 
         self.outputs['vertices'].label = 'Vertices'
         self.outputs['vertices'].custom_draw = 'draw_vertices_out_socket'
         self.outputs['edges'].label = 'Edges'
         self.outputs['polygons'].label = 'Polygons'
-        # self.outputs['offsets'].label = 'Offsets'
-        # self.outputs['altitudes'].label = 'Altitudes'
+        #self.outputs['offsets'].label = 'Offsets'
+        #self.outputs['altitudes'].label = 'Altitudes'
         self.outputs['failed_contours_vertices'].label = 'No wrong contours verts'
         self.outputs['failed_contours_vertices'].custom_draw = 'draw_failed_contours_vertices_out_socket'
 
@@ -541,29 +558,16 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
             return
         
         inputs = self.inputs
-        _Vertices  = inputs['vertices'].sv_get(default=[[]], deepcopy=False)
-        Vertices3  = ensure_nesting_level(_Vertices, 3)
-        _Edges     = inputs['edges'].sv_get(default=[[]], deepcopy=False)
-        Edges3     = ensure_nesting_level(_Edges, 3)
-        _Faces     = inputs['polygons'].sv_get(default=[[]], deepcopy=False)
-        Faces3     = ensure_nesting_level(_Faces, 3)
-
-        # _ss_offsets  = inputs['ss_offsets'].sv_get(default=[[self.ss_offset1]], deepcopy=False)
-        # ss_offsets2  = ensure_nesting_level(_ss_offsets, 2)
-        # _ss_altitudes  = inputs['ss_altitudes'].sv_get(default=[[self.ss_altitude1]], deepcopy=False)
-        # ss_altitudes2  = ensure_nesting_level(_ss_altitudes, 2)
-
-        _ss_angles  = inputs['ss_angles'].sv_get(default=[[self.ss_angles]], deepcopy=False)
-        if self.angles_mode=='OBJECTS':
-            ss_angles2  = ensure_nesting_level(_ss_angles, 2)
-        else:
-            ss_angles3  = ensure_nesting_level(_ss_angles, 3)
-
-        _ss_restrict_heights  = inputs['ss_restrict_heights'].sv_get(default=[[self.ss_restrict_height1]], deepcopy=False)
-        ss_restrict_heights  = ensure_nesting_level(_ss_restrict_heights, 2)
-        _ss_heights  = inputs['ss_height'].sv_get(default=[[self.ss_height]], deepcopy=False)
-        ss_heights1  = ensure_nesting_level(_ss_heights, 2)
-
+        _Vertices       = inputs['vertices'].sv_get(default=[[]], deepcopy=False)
+        Vertices3       = ensure_nesting_level(_Vertices, 3)
+        _Edges          = inputs['edges'].sv_get(default=[[]], deepcopy=False)
+        Edges3          = ensure_nesting_level(_Edges, 3)
+        _Faces          = inputs['polygons'].sv_get(default=[[]], deepcopy=False)
+        Faces3          = ensure_nesting_level(_Faces, 3)
+        _ss_offsets     = inputs['ss_offsets'].sv_get(default=[[self.ss_offset1]], deepcopy=False)
+        ss_offsets2     = ensure_nesting_level(_ss_offsets, 2)
+        _ss_altitudes   = inputs['ss_altitudes'].sv_get(default=[[self.ss_altitude1]], deepcopy=False)
+        ss_altitudes2   = ensure_nesting_level(_ss_altitudes, 2)
 
         # selecte shape mode in property
         ss_shapes_mode1 = [I for I, shapes_modes in enumerate(self.ss_shapes_modes) if shapes_modes[0] == self.ss_shapes_mode1]
@@ -587,12 +591,12 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
         res_boundaries_verts = []
         res_edges = []
         res_faces = []
-        res_offsets = []   # what offset used for results objects
-        res_altitudes = [] # what altitude used for results objects
+        #res_offsets = []   # what offset used for results objects
+        #res_altitudes = [] # what altitude used for results objects
 
         objects_data = dict()
-        objects_offsets_of_boundaries = []
-        objects_area_boundaries = []
+        #objects_offsets_of_boundaries = []
+        #objects_area_boundaries = []
 
         contours_failed_at_all = []
         params = zip_long_repeat(Vertices3, Edges3, Faces3)
@@ -628,50 +632,32 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
             if mask==True:
                 continue
 
-            # if self.offset_mode=='OBJECT_ALL_OFFSETS':
-            #     if I<=len(ss_offsets2)-1:
-            #         ss_offsets = ss_offsets2[I]
-            #     else:
-            #         ss_offsets = ss_offsets2[-1]
-            #     pass
-            # elif self.offset_mode=='OBJECT_ONE_OFFSET':
-            #     if I<=len(ss_offsets2[0])-1:
-            #         ss_offsets = [ss_offsets2[0][I]]
-            #     else:
-            #         ss_offsets = [ss_offsets2[0][-1]]
-            #     pass
-            
-            # if self.altitude_mode=='OBJECT_ALL_ALTITUDES':
-            #     if I<=len(ss_altitudes2)-1:
-            #         ss_altitudes = ss_altitudes2[I]
-            #     else:
-            #         ss_altitudes = ss_altitudes2[-1]
-            #     pass
-            # elif self.altitude_mode=='OBJECT_ONE_ALTITUDE':
-            #     if I<=len(ss_altitudes2[0])-1:
-            #         ss_altitudes = [ss_altitudes2[0][I]]
-            #     else:
-            #         ss_altitudes = [ss_altitudes2[0][-1]]
-            #     pass
-            
-            if I<=len(ss_restrict_heights[0])-1:
-                ss_restrict_height1 = ss_restrict_heights[0][I]
-            else:
-                ss_restrict_height1 = ss_restrict_heights[0][-1]
-
-            if I<=len(ss_heights1[0])-1:
-                ss_height1 = ss_heights1[0][I]
-            else:
-                ss_height1 = ss_heights1[0][-1]
-
-            if self.angles_mode=='OBJECTS':
-                if I<=len(ss_angles2[0])-1:
-                    ss_angle = ss_angles2[0][I]
+            if self.offset_mode=='OBJECT_ALL_OFFSETS':
+                if I<=len(ss_offsets2)-1:
+                    ss_offsets = ss_offsets2[I]
                 else:
-                    ss_angle = ss_angles2[0][-1]
-            else:
-                raise f'Stright Skeleton angles_mode={self.angles_mode} not reailized'
-
+                    ss_offsets = ss_offsets2[-1]
+                pass
+            elif self.offset_mode=='OBJECT_ONE_OFFSET':
+                if I<=len(ss_offsets2[0])-1:
+                    ss_offsets = [ss_offsets2[0][I]]
+                else:
+                    ss_offsets = [ss_offsets2[0][-1]]
+                pass
+            
+            if self.altitude_mode=='OBJECT_ALL_ALTITUDES':
+                if I<=len(ss_altitudes2)-1:
+                    ss_altitudes = ss_altitudes2[I]
+                else:
+                    ss_altitudes = ss_altitudes2[-1]
+                pass
+            elif self.altitude_mode=='OBJECT_ONE_ALTITUDE':
+                if I<=len(ss_altitudes2[0])-1:
+                    ss_altitudes = [ss_altitudes2[0][I]]
+                else:
+                    ss_altitudes = [ss_altitudes2[0][-1]]
+                pass
+            
             if _shapes_modes[I]<0 or len(self.ss_shapes_modes) < _shapes_modes[I]:
                 raise Exception(f"unknown Shapes mode value: '{_shapes_modes[I]}'. Allowed values [{allowed_shapes_modes}]")
             
@@ -679,35 +665,62 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
 
             # separate objects of loose parts (objects can has islands. Every island have to be separated)
             if not faces_i or not faces_i[0]:
-                raise Exception(f"Error: Object {I} has no faces. Extrusion is not possible. Objects should be flat.")
+                raise Exception(f"Error: Object {I} has no faces. Straight Skeleton Offset is not possible. Objects should be flat.")
             
+            # Попытка разделить graph mesh  на контуры через networkx не особо помогла. Производительность не улучшилась, но библиотека всё равно интересная
+            # time_1_1 = time()
+            # try:
+            #     # Разделить объект на контуры:
+            #     bm = bmesh_from_pydata(verts_i, edges_i, faces_i, normal_update=True)
+            #     bm.edges.ensure_lookup_table()
+            #     edges = [[e.verts[0].index, e.verts[1].index] for e in bm.edges]
+            #     object_I_contours_edges, _, _ = Edges.process(bm, "Boundary", edges)
+            #     object_I_verts = separate_edges(verts_i, object_I_contours_edges)
+            #     bm.free()
+            # except Exception as ex:
+            #     # Keep shape to show as errors in the future
+            #     contours_failed_at_all.append(verts_i)
+            #     continue
+            # time_1_1 = time()-time_1_1
+            # print(f'object {I} separates parts: {time_1_1}, contours: {len(object_I_verts)}')
+
+            time_4_1 = time()
+            # object_I_planes_verts, object_I_planes_faces, _ = separate_edges(verts_i, edges_i)
             object_I_planes_verts, object_I_planes_faces, _ = separate_loose_mesh(verts_i, faces_i)
 
             for IJ in range(len(object_I_planes_verts)):
                 object_I_plane_IJ_verts, object_I_plane_IJ_faces = object_I_planes_verts[IJ], object_I_planes_faces[IJ]
 
+                time_2_1 = time()
                 try:
                     bm = bmesh_from_pydata(object_I_plane_IJ_verts, None, object_I_plane_IJ_faces, normal_update=True)
                     bm.edges.ensure_lookup_table()
                     edges = [[e.verts[0].index, e.verts[1].index] for e in bm.edges]
                     object_I_plane_IJ_contours_edges, _, _ = Edges.process(bm, "Boundary", edges)
                     bm.free()
+
                 except Exception as ex:
                     # Keep shape to show as errors in the future
                     contours_failed_at_all.append(object_I_plane_IJ_verts)
                     continue
+                if self.verbose_messages_while_process==True:
+                    time_2_1 = time()-time_2_1
+                    print(f'object {I}, part {IJ} calc baunadries: {time_2_1}')
 
                 if not object_I_plane_IJ_contours_edges:
                     raise Exception(f"Error: Object {I} has no boundaries. Extrusion is not possible. Objects should be flat.")
                 # separate contours of every island
+                #object_I_plane_IJ_contours_verts, edges_boundaries, _ = separate_edges(object_I_plane_IJ_verts, object_I_plane_IJ_contours_edges)
                 object_I_plane_IJ_contours_verts, edges_boundaries, _ = separate_loose_mesh(object_I_plane_IJ_verts, object_I_plane_IJ_contours_edges)
 
                 object_I_plane_IJ_contours = []
                 object_area_boundaries = []
-                objects_angles_of_boundary = []
-                objects_offsets_of_boundary = []
+                #objects_offsets_of_boundary = []
                 failed_contours_vertices = []
                 areas = []
+                
+                time_3_1 = time()
+                
                 for IJK in range(len(object_I_plane_IJ_contours_verts)):
                     object_I_plane_IJ_contour_IJK_verts, object_I_plane_IJ_contour_IJK_edges = object_I_plane_IJ_contours_verts[len(object_I_plane_IJ_contours_verts)-1-IJK], edges_boundaries[len(object_I_plane_IJ_contours_verts)-1-IJK]
                     object_I_plane_IJ_contour_IJK_verts_sorted = vertices_sort_by_edges(object_I_plane_IJ_contour_IJK_verts, object_I_plane_IJ_contour_IJK_edges)
@@ -716,23 +729,28 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
                     areas.append(area[0])
                     object_I_plane_IJ_contours.append(object_I_plane_IJ_contour_IJK_verts_sorted)
                     object_area_boundaries.append({"area":area, "object_idx":I, "object_boundaries":object_I_plane_IJ_contour_IJK_verts_sorted})
-                    objects_angles_of_boundary.append( [ss_angle*180/math.pi,]*len(object_I_plane_IJ_contour_IJK_verts_sorted) )
                     pass
-                srt = sort_together([areas, object_I_plane_IJ_contours, objects_angles_of_boundary])
+                srt = sort_together([areas, object_I_plane_IJ_contours, ])
                 object_I_plane_IJ_contours_sorted_by_area = list(reversed(srt[1]))  # First contour is outer boundary - another is holes
-                object_I_plane_IJ_angles_sorted_by_area = list(reversed(srt[2]))  # First contour is outer boundary - another is holes
+
+                if self.verbose_messages_while_process==True:
+                    time_3_1 = time()-time_3_1
+                    print(f'object {I}, part {IJ} process baundaries: {time_3_1}')
 
                 if shapes_mode_1 in [ 'ORIGINAL_BOUNDARIES', 'EXCLUDE_HOLES', 'INVERT_HOLES', ]: # and len(object_boundaries_sorted_by_area)>1:
                     shapes_mode_1_id = [info for info in self.ss_shapes_modes if info[0]==shapes_mode_1][0][4]
                     if I not in objects_data:
-                        objects_data[I] = {"idx": I, 'restrict_height': ss_restrict_height1, 'height': ss_height1, 'planes':[], 'angles':[], 'shape_mode':shapes_mode_1_id}
+                        objects_data[I] = {"idx": I, 'offsets': ss_offsets, 'altitudes': ss_altitudes, 'planes':[], 'shape_mode':shapes_mode_1_id}
                     objects_data[I]['planes'].append(object_I_plane_IJ_contours_sorted_by_area)
-                    objects_data[I]['angles'].append(object_I_plane_IJ_angles_sorted_by_area)
                     pass
                 else:
                     raise Exception(f"unknown Shapes mode value: '{_shapes_modes[I]}'. Allowed values {allowed_shapes_modes}")
                 pass
             pass
+
+            if self.verbose_messages_while_process==True:
+                time_4_1 = time()-time_4_1
+                print(f'object calc baunadries general time: {time_4_1}')
         pass
 
         errors_vertices = []
@@ -740,75 +758,55 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
             lst_errors = []
             was_errors = False
             
-            # if(self.res_type=='CONTOURS'):
-            #     res_type=0
-            # elif(self.res_type=='FACES'):
-            #     res_type=1
-            # else:
-            #     raise Exception(f"Unknown res_type={self.res_type}. Allowed only 'CONTOURS' or 'FACES'.")
+            if(self.res_type=='CONTOURS'):
+                res_type=0
+            elif(self.res_type=='FACES'):
+                res_type=1
+            else:
+                raise Exception(f"Unknown res_type={self.res_type}. Allowed only 'CONTOURS' or 'FACES'.")
 
-            join_mode_id = [info for info in self.join_modes if info[0]==self.join_mode][0][4]
+            source_objects_join_mode_id = [info for info in self.source_objects_join_modes if info[0]==self.source_objects_join_mode][0][4]
+            results_join_mode_id = [info for info in self.results_join_modes if info[0]==self.results_join_mode][0][4]
             data = {
                 'objects' : [],
-                'force_z_zero': self.force_z_zero, 
-                #'res_type': res_type, 
-                'only_tests_for_valid': self.only_tests_for_valid, 
-                'join_mode': join_mode_id, # 0 - split, 1 - keep, 2 - merge all meshes
-                'verbose' : self.verbose_messages_while_process,
+                'force_z_zero'          : self.force_z_zero, 
+                'res_type': res_type, 
+                'source_objects_join_mode'             : source_objects_join_mode_id,
+                'results_join_mode'             : results_join_mode_id,
+                'only_tests_for_valid'  : self.only_tests_for_valid, 
+                'verbose'               : self.verbose_messages_while_process,
             }
             
             for I in range(len(objects_data)):
                 objects_data_I       = objects_data[I]
                 shape_mode_I         = objects_data_I["shape_mode"]
-                angles_I             = objects_data_I["angles"]
-                # offsets_I            = objects_data_I["offsets"]
-                # altitudes_I          = objects_data_I["altitudes"]
-                restrict_height_I    = objects_data_I["restrict_height"]
-                ss_height_I          = objects_data_I["height"]
-                # offsets = []
-                # altitudes = []
-                # for offset_index, offset1 in enumerate(offsets_I):
-                #     if offset_index<=len(ss_altitudes)-1:
-                #         ss_altitude1 = altitudes_I[offset_index]
-                #     else:
-                #         ss_altitude1 = altitudes_I[-1]
-                #     offsets.append(offset1)
-                #     altitudes.append(ss_altitude1)
+                offsets_I            = objects_data_I["offsets"]
+                altitudes_I          = objects_data_I["altitudes"]
+                offsets = []
+                altitudes = []
+                for offset_index, offset1 in enumerate(offsets_I):
+                    if offset_index<=len(ss_altitudes)-1:
+                        ss_altitude1 = altitudes_I[offset_index]
+                    else:
+                        ss_altitude1 = altitudes_I[-1]
+                    offsets.append(offset1)
+                    altitudes.append(ss_altitude1)
                 
-                # if len(offsets)>0:
-                data['objects'].append( {
-                    'object_id':objects_data_I['idx'],
-                    'shape_mode' : shape_mode_I,
-                    #'polygon_id':I, 
-                    # 'offsets': offsets,
-                    # 'altitudes'       : altitudes,
-                    'restrict_height' : objects_data_I["restrict_height"],
-                    'height'          : objects_data_I["height"],
-                    'planes'          : objects_data_I["planes"], # I is not wrong, boundary1 (array of contours) - plane
-                    'angles'          : objects_data_I["angles"], # I is not wrong, boundary1 (array of contours) - plane
-                } )
+                if len(offsets)>0:
+                    data['objects'].append( {
+                        'object_id':objects_data_I['idx'],
+                        'shape_mode' : shape_mode_I,
+                        #'polygon_id':I, 
+                        'offsets': offsets,
+                        'altitudes': altitudes,
+                        'planes' : objects_data_I["planes"], # I is not wrong, boundary1 (array of contours) - plane
+                    } )
 
             # run all skeletons in Threads
             #data_processed = list( executor.map(parallel_straight_skeleton_2d_offset, data))
-            data_processed = pySVCGAL_straight_skeleton_2d_extrude(data)
-            faces_delta = 0
-            
-            object_verts_merge = []
-            object_edges_merge = []
-            object_faces_merge = []
-
-            objects_verts_keep = dict()
-            objects_edges_keep = dict()
-            objects_faces_keep = dict()
-
+            data_processed = pySVCGAL_straight_skeleton_2d_offset(data)
             lst_errors1 = []
-
-            valid_offsets_merge = []
-            valid_altitudes_merge = []
-
-            valid_offsets_keep = dict()
-            valid_altitudes_keep = dict()
-
+            
             for data1 in data_processed['source_objects_errors']: 
                 object_index = data1["object_index"]
                 error_vertices_object1 = []
@@ -821,100 +819,20 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
                 errors_vertices.append(error_vertices_object1)
                 pass
 
-
             for data1 in data_processed['objects']:
-                # object_index = data1["object_index"]
-                # error_vertices_object1 = []
-                # if(len(data1['vertices_of_errors'])>0):
-                #     error_vertices_object1 = data1['vertices_of_errors']
-                #     if self.verbose_messages_while_process==True:
-                #         print(f'Object {object_index} has errors:')
-                #         for s in data1['descriptions_per_errors']:
-                #             print(f'    {s}')
-                # errors_vertices.append(error_vertices_object1)
-
                 # Даже если была ошибка, то проверить, может есть возможность отобразить хоть какие-то данные? Тем более, если ошибки не было!
                 if self.only_tests_for_valid==True:
                     # no result, no output
                     pass
                 else:
-                    # # Нельзя пропускать, иначе не будет соответствия исходных объектов и результирующих если по кому-то не получено результатов!!!
-                    # # if len(data1["vertices"])==0:
-                    # #     pass
-                    # # else:
-                    #     if self.join_mode=='SPLIT':
-                    #         # res_offsets.append( [ data1['offset'] ] )
-                    #         # res_altitudes.append( [ data1['altitude'] ] )
-
-                    #         res_verts.append( data1['vertices'] )
-                    #         res_edges.append( data1['edges']    )
-                    #         res_faces.append( data1['faces']    )
-                    #         pass
-                    #     elif self.join_mode=='KEEP':
-                    #         object_index = data1['object_index']
-                    #         if object_index not in objects_verts_keep:
-                    #             objects_verts_keep[object_index] = []
-                    #             objects_edges_keep[object_index] = []
-                    #             objects_faces_keep[object_index] = []
-
-                    #             valid_offsets_keep  [object_index] = []
-                    #             valid_altitudes_keep[object_index] = []
-
-                    #             pass
-                    #         objects_verts_keep[object_index].extend( data1['vertices'] )
-                    #         objects_edges_keep[object_index].extend( data1['edges'] )
-                    #         objects_faces_keep[object_index].extend( data1['faces'] )
-
-                    #         # valid_offsets_keep  [object_id].append(data1['offset'])
-                    #         # valid_altitudes_keep[object_id].append(data1['altitude'])
-
-                    #         pass
-                    #     elif self.join_mode=='MERGE':
-                    #         object_verts_merge.extend(data1['vertices'])
-                    #         object_edges_merge.extend( [ list(map(lambda n: n+faces_delta, face)) for face in data1['edges'] ] )
-                    #         object_faces_merge.extend( [ list(map(lambda n: n+faces_delta, face)) for face in data1['faces'] ] )
-                    #         faces_delta+=len(data1['vertices'])
-
-                    pass # теперь библиотека C++ возвращает результат в нужном join_mode
                     object_index = data1['object_index']
                     res_verts.append(data1['vertices'])
                     res_edges.append(data1['edges'])
                     res_faces.append(data1['faces'])
-                    # if object_index not in objects_verts_keep:
-                    #     objects_verts_keep[object_index] = []
-                    #     objects_edges_keep[object_index] = []
-                    #     objects_faces_keep[object_index] = []
-                    #     valid_offsets_keep  [object_index] = []
-                    #     valid_altitudes_keep[object_index] = []
-                    #     pass
-
-                    # objects_verts_keep[object_index].extend( data1['vertices'] )
-                    # objects_edges_keep[object_index].extend( data1['edges'] )
-                    # objects_faces_keep[object_index].extend( data1['faces'] )
                     pass
-
                 pass
             lst_errors.extend(lst_errors1)
-            # if self.join_mode=='MERGE':
-            #     res_verts.append(object_verts_merge)
-            #     res_edges.append(object_edges_merge)
-            #     res_faces.append(object_faces_merge)
-
-            #     res_offsets.append(valid_offsets_merge)
-            #     res_altitudes.append(valid_altitudes_merge)
-
-            #     pass
-
-            # elif self.join_mode=='KEEP':
-            #     for KEY in objects_verts_keep:
-            #         res_verts.append(objects_verts_keep[KEY])
-            #         res_edges.append(objects_edges_keep[KEY])
-            #         res_faces.append(objects_faces_keep[KEY])
-
-            #         res_offsets.append(valid_offsets_keep[KEY])
-            #         res_altitudes.append(valid_altitudes_keep[KEY])
-            # else:
-            #     pass
+            
             if len(contours_failed_at_all)>0:
                 failed_contours_vertices.extend(contours_failed_at_all)
             pass
@@ -961,7 +879,7 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
             # txt_verts  = "\n".join(lines_verts)
             # txt_angles = "\n".join(lines_angles)
 
-            # print(f"stright skeleton node write to file")
+            # print(f"straight skeleton node write to file")
             # with open(file_name_dat, "w") as file:
             #     file.write(txt_verts)
             #     print(f'Записаны вершины {len(lines_verts)-1}: {file_name_dat}')
@@ -973,8 +891,8 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
         self.outputs['vertices'].sv_set(res_verts)
         self.outputs['edges'].sv_set(res_edges)
         self.outputs['polygons'].sv_set(res_faces)
-        # self.outputs['offsets'].sv_set(res_offsets)
-        # self.outputs['altitudes'].sv_set(res_altitudes)
+        #self.outputs['offsets'].sv_set(res_offsets)
+        #self.outputs['altitudes'].sv_set(res_altitudes)
         # Test any data in errors
         is_errors = False
         for elem in errors_vertices:
@@ -996,5 +914,5 @@ class SvStraightSkeleton2DExtrude(ModifierLiteNode, SverchCustomTreeNode, bpy.ty
         print("file .dat saved")
         pass
 
-classes = [SvStraightSkeleton2DExtrude,]
+classes = [SvStraightSkeleton2DOffset,]
 register, unregister = bpy.utils.register_classes_factory(classes)
